Clean up CUDA Event dependencies
Move the CUDA Event specialization include out of the generic Event header and into the CUDA declaration header. Let the CUDA Event default constructor delegate through the execution-space constructor.

Inline the remaining portable Event test functors into TestEvent.hpp and remove the now-unused helper header.
