--- core/unit_test/TestEventHelpers.hpp
@@ -1,54 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-// SPDX-FileCopyrightText: Copyright Contributors to the Kokkos project
-
-#ifndef KOKKOS_TEST_EVENT_HELPERS_HPP
-#define KOKKOS_TEST_EVENT_HELPERS_HPP
-
-#include <Kokkos_Macros.hpp>
-#ifdef KOKKOS_ENABLE_EXPERIMENTAL_CXX20_MODULES
-import kokkos.core;
-#else
-#include <Kokkos_Core.hpp>
-#endif
-
-#include <concepts>
-
-// Functors shared between the portable Event tests (TestEvent.hpp) and the
-// CUDA-specific Event tests (cuda/TestCuda_CudaEvent.cpp).
-// NVCC cannot handle extended __host__ __device__ lambdas inside the private
-// TestBody() generated by the TEST() macro, so named functors are required.
-
-namespace Test {
-
-template <class ViewType>
-struct FillFunctor {
-  ViewType data;
-  template <std::integral T>
-  KOKKOS_FUNCTION void operator()(T i) const {
-    data(i) = i + 1;
-  }
-};
-
-template <class ViewType>
-struct ProduceFunctor {
-  ViewType data;
-  template <std::integral T>
-  KOKKOS_FUNCTION void operator()(T i) const {
-    data(i) = i * 2;
-  }
-};
-
-template <class ViewType, class ResultType>
-struct ConsumeFunctor {
-  ViewType data;
-  ResultType result;
-  template <std::integral T>
-  KOKKOS_FUNCTION void operator()(T i) const {
-    Kokkos::atomic_add(&result(),
-                       static_cast<typename ResultType::value_type>(data(i)));
-  }
-};
-
-}  // namespace Test
-
-#endif  // KOKKOS_TEST_EVENT_HELPERS_HPP
